// frontend/src/proto/user_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts,import_extension=ts"
// @generated from file user.proto (package musicclub.user, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file user.proto.
 */
export const file_user: GenFile = /*@__PURE__*/
  fileDesc("Cgp1c2VyLnByb3RvEg5tdXNpY2NsdWIudXNlciJjCgRVc2VyEgoKAmlkGAEgASgJEhQKDGRpc3BsYXlfbmFtZRgCIAEoCRIQCgh1c2VybmFtZRgDIAEoCRISCgphdmF0YXJfdXJsGAQgASgJEhMKC3RlbGVncmFtX2lkGAUgASgEQhxaGm11c2ljY2x1YmJvdC9iYWNrZW5kL3Byb3RvYgZwcm90bzM");

/**
 * Minimal user info for displaying assignments and ownership.
 *
 * @generated from message musicclub.user.User
 */
export type User = Message<"musicclub.user.User"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string display_name = 2;
   */
  displayName: string;

  /**
   * @generated from field: string username = 3;
   */
  username: string;

  /**
   * @generated from field: string avatar_url = 4;
   */
  avatarUrl: string;

  /**
   * @generated from field: uint64 telegram_id = 5;
   */
  telegramId: bigint;
};

/**
 * Describes the message musicclub.user.User.
 * Use `create(UserSchema)` to create a new message.
 */
export const UserSchema: GenMessage<User> = /*@__PURE__*/
  messageDesc(file_user, 0);